Remove PartialEq, Eq, PartialOrd, Ord implementations
The problem is that regexes cannot be compared, so we remove the
implementations for PartialEq, Eq, PartialOrd and Ord here.

The only benefit which came from these implementations was that we can
use the BinaryHeap instead of a Vec for storing these. The performance
impact should be rather minimal here, as we only have a few (like,...
less than 100) elements in the Vec... so not much difference between
them.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>